# Batch laguerre series computation in wigner